+++ fltk-sys/src/window.rs
@@ -2980,6 +2980,9 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Window_default_cursor(self_: *mut Fl_Window, cursor: libc::c_int);
+}
+extern "C" {
     pub fn Fl_Window_find_by_handle(handle: *mut libc::c_void) -> *mut Fl_Window;
 }
 extern "C" {
@@ -3414,6 +3417,9 @@ extern "C" {
         hot_y: libc::c_int,
     );
 }
+extern "C" {
+    pub fn Fl_Single_Window_default_cursor(self_: *mut Fl_Single_Window, cursor: libc::c_int);
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Double_Window {
@@ -3830,6 +3836,9 @@ extern "C" {
         hot_y: libc::c_int,
     );
 }
+extern "C" {
+    pub fn Fl_Double_Window_default_cursor(self_: *mut Fl_Double_Window, cursor: libc::c_int);
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Menu_Window {
@@ -4239,6 +4248,9 @@ extern "C" {
         hot_x: libc::c_int,
         hot_y: libc::c_int,
     );
+}
+extern "C" {
+    pub fn Fl_Menu_Window_default_cursor(self_: *mut Fl_Menu_Window, cursor: libc::c_int);
 }
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -4676,6 +4688,9 @@ extern "C" {
         hot_y: libc::c_int,
     );
 }
+extern "C" {
+    pub fn Fl_Overlay_Window_default_cursor(self_: *mut Fl_Overlay_Window, cursor: libc::c_int);
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Gl_Window {
@@ -5075,6 +5090,9 @@ extern "C" {
         hot_x: libc::c_int,
         hot_y: libc::c_int,
     );
+}
+extern "C" {
+    pub fn Fl_Gl_Window_default_cursor(self_: *mut Fl_Gl_Window, cursor: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Gl_Window_flush(self_: *mut Fl_Gl_Window);
@@ -5552,6 +5570,9 @@ extern "C" {
         hot_x: libc::c_int,
         hot_y: libc::c_int,
     );
+}
+extern "C" {
+    pub fn Fl_Glut_Window_default_cursor(self_: *mut Fl_Glut_Window, cursor: libc::c_int);
 }
 extern "C" {
     pub fn Fl_Glut_Window_flush(self_: *mut Fl_Glut_Window);